Feat: Separate config files per abstract solver manager
