+++ source/isaaclab_newton/isaaclab_newton/physics/__init__.pyi
@@ -21,17 +21,17 @@ __all__ = [
 ]
 
 from .featherstone_manager import NewtonFeatherstoneManager
+from .featherstone_manager_cfg import FeatherstoneSolverCfg
 from .kamino_manager import NewtonKaminoManager
+from .kamino_manager_cfg import KaminoSolverCfg
 from .mjwarp_manager import NewtonMJWarpManager
+from .mjwarp_manager_cfg import MJWarpSolverCfg
 from .newton_collision_cfg import HydroelasticSDFCfg, NewtonCollisionPipelineCfg
 from .newton_manager import NewtonManager
 from .newton_manager_cfg import (
-    FeatherstoneSolverCfg,
-    KaminoSolverCfg,
-    MJWarpSolverCfg,
     NewtonCfg,
     NewtonShapeCfg,
     NewtonSolverCfg,
-    XPBDSolverCfg,
 )
 from .xpbd_manager import NewtonXPBDManager
+from .xpbd_manager_cfg import XPBDSolverCfg
